Add files via upload
Data Collection Methodology: 
This data is collected by Boston Athletic Association, the official organizer of the Boston Marathon. It tracked each year record of  the official finishing time of the Men’s Open Division winner. The data is recorded annually on the marathon day, and stored as an annual time series with times in minutes.

Source: Publicly available historical records on the Boston Athletic Association’s website: https://www.baa.org/races/boston-marathon/results/champions

Why does this data set intrigue you?
The reason I chose this dataset is that it captures more than a century of human athletic performance in the Boston Marathon, one of the most iconic races in the world. 
What makes it really cool is that we can combine other historical data, such as World Wars, and extreme weather, and see the impacts, sometimes caused surprising spikes or drops in the winning times. And there’s also the fun part of using this data for long-term forecasting. We also can model whether winning times are getting close to a human performance limit and even try to predict when we might see the next record-breaking run.